Android QThrad::create problem
Work around, due to this good, on the Android platform. Because of the
dependency to pthread which is compiled inside LibC in Android, the test
module is not built. And as a consequence, support for cxx11_future is
disabled on the Android platform in Qt versions up to 6.0. https://
github.com/qt/qtbase/commit/6b488d89507e4d252b05225a203183b5e83e92ac